API update

+++ stepmania/src/arch/Sound/ALSA9Helpers.cpp
@@ -303,7 +303,7 @@ bool Alsa9Buf::Recover( int r )
 	return false;
 }
 
-int Alsa9Buf::GetPosition() const
+int64_t Alsa9Buf::GetPosition() const
 {
 	if( dsnd_pcm_state(pcm) == SND_PCM_STATE_PREPARED )
 		return last_cursor_pos;
